requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
backports-asyncio-runner==1.2.0
blake3==1.0.8
bzip2==1.0.8
ca-certificates==2025.9.9
cachetools==5.5.2
cbor2==5.7.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
compressed-tensors==0.11.0
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
datasets==4.2.0
debugpy==1.8.16
decorator==5.2.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
dotenv==0.9.9
einops==0.8.1
elevenlabs==2.21.0
email-validator==2.3.0
eval-type-backport==0.2.2
exceptiongroup==1.3.0
execnet==2.1.1
executing==2.2.1
expat==2.7.1
fastapi==0.121.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
filelock==3.20.0
fonttools==4.60.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.9.0
gguf==0.17.1
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.124.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-genai==1.47.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.3
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.3.0
interegular==0.3.3
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
krb5==1.21.3
lark==1.2.2
lazy-loader==0.4
liac-arff==2.5.0
libcxx==20.1.8
libedit==3.1.20250104
libffi==3.4.4
librosa==0.11.0
libsodium==1.0.20
libzlib==1.3.1
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.18
mistral-common==1.8.5
mlx==0.29.4
mlx-lm==0.28.3
mlx-metal==0.29.4
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
multiprocess==0.70.16
narwhals==2.10.0
ncurses==6.5
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
numba==0.61.2
numpy==2.2.6
openai==2.4.0
openai-harmony==0.0.8
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openml==0.15.1
openssl==3.5.3
outlines-core==0.2.11
packaging==25.0
pandas==2.3.3
parso==0.8.5
partial-json-parser==0.2.1.1.post7
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==25.2
platformdirs==4.4.0
plotly==6.3.1
pluggy==1.6.0
pooch==1.8.2
prometheus-client==0.23.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-extra-types==2.10.6
pydub==0.25.1
pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-xdist==3.8.0
python==3.10.18
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-io==0.3
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
ray==2.51.1
readline==8.3
redis==7.0.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.6.2
safetytooling==0.0.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.1
sentry-sdk==2.43.0
setproctitle==1.3.7
setuptools==78.1.1
shapely==2.1.2
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
sqlite==3.50.2
stack_data==0.6.3
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tk==8.6.15
together==1.5.29
tokenizers==0.22.1
tomli==2.3.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.11.0
wandb==0.22.3
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
xgrammar==0.1.25
xmltodict==1.0.2
xxhash==3.6.0
xz==5.6.4
yarl==1.22.0
zeromq==4.3.5
zipp==3.23.0
zlib==1.3.1
